refactor: 전체 로직 chess 패키지 안으로 이동
