--- src/main/java/Application.java
@@ -1,5 +1,0 @@
-public class Application {
-    public static void main(String[] args) {
-
-    }
-}
