+++ src/main/java/chess/Application.java
@@ -1,0 +1,7 @@
+package chess;
+
+public class Application {
+    public static void main(String[] args) {
+
+    }
+}
