+++ src/main/java/chess/domain/Player.java
@@ -1,0 +1,15 @@
+package chess.domain;
+
+import chess.domain.piece.Color;
+
+public class Player {
+    private final Color color;
+
+    public Player(Color color) {
+        this.color = color;
+    }
+
+    public Color getColor() {
+        return color;
+    }
+}
