+++ src/main/java/chess/domain/piece/Color.java
@@ -1,0 +1,13 @@
+package chess.domain.piece;
+
+public enum Color {
+    BLACK,
+    WHITE;
+
+    public int moveUnit() {
+        if (this == BLACK) {
+            return -1;
+        }
+        return 1;
+    }
+}
