+++ src/main/java/chess/view/InputView.java
@@ -1,0 +1,4 @@
+package chess.view;
+
+public class InputView {
+}
