+++ src/main/java/chess/view/OutputView.java
@@ -1,0 +1,4 @@
+package chess.view;
+
+public class OutputView {
+}
