--- src/main/java/domain/Player.java
@@ -1,15 +1,0 @@
-package domain;
-
-import domain.piece.Color;
-
-public class Player {
-    private final Color color;
-
-    public Player(Color color) {
-        this.color = color;
-    }
-
-    public Color getColor() {
-        return color;
-    }
-}
