--- src/main/java/domain/piece/Color.java
@@ -1,13 +1,0 @@
-package domain.piece;
-
-public enum Color {
-    BLACK,
-    WHITE;
-
-    public int moveUnit() {
-        if (this == BLACK) {
-            return -1;
-        }
-        return 1;
-    }
-}
