--- src/main/java/view/InputView.java
@@ -1,4 +1,0 @@
-package view;
-
-public class InputView {
-}
