--- src/main/java/view/OutputView.java
@@ -1,4 +1,0 @@
-package view;
-
-public class OutputView {
-}
