+++ src/test/java/chess/domain/piece/KnightTest.java
@@ -1,0 +1,42 @@
+package chess.domain.piece;
+
+import chess.domain.position.Position;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.Arguments;
+import org.junit.jupiter.params.provider.MethodSource;
+
+import java.util.stream.Stream;
+
+import static org.assertj.core.api.Assertions.assertThatCode;
+import static org.junit.jupiter.api.Assertions.assertTrue;
+
+class KnightTest {
+    private static Stream<Arguments> destinations() {
+        return Stream.of(
+                Arguments.of(Position.from("b7")),
+                Arguments.of(Position.from("d7")),
+                Arguments.of(Position.from("e6")),
+                Arguments.of(Position.from("e4")),
+                Arguments.of(Position.from("d3")),
+                Arguments.of(Position.from("b3")),
+                Arguments.of(Position.from("a4")),
+                Arguments.of(Position.from("a6"))
+        );
+    }
+
+    @ParameterizedTest
+    @MethodSource("destinations")
+    void move(Position position) {
+        Knight knight = new Knight(Color.BLACK, Position.from("c5"));
+        knight.move(position, new Pieces());
+        assertTrue(knight.hasPosition(position));
+    }
+
+    @Test
+    void cornerMove() {
+        Knight knight = new Knight(Color.BLACK, Position.from("a5"));
+        assertThatCode(() -> knight.move(Position.from("c4"), new Pieces()))
+                .doesNotThrowAnyException();
+    }
+}
