+++ src/test/java/chess/domain/position/RowTest.java
@@ -1,0 +1,43 @@
+package chess.domain.position;
+
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.Arguments;
+import org.junit.jupiter.params.provider.MethodSource;
+
+import java.util.stream.Stream;
+
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.Assertions.assertThatThrownBy;
+
+class RowTest {
+
+    private static Stream<Arguments> rows() {
+        return Stream.of(
+                Arguments.of(-5, Row.ONE),
+                Arguments.of(-4, Row.TWO),
+                Arguments.of(0, Row.SIX),
+                Arguments.of(2, Row.EIGHT)
+        );
+    }
+
+    @ParameterizedTest
+    @MethodSource("rows")
+    void moveBy(int value, Row expected) {
+        Row row = Row.SIX;
+        assertThat(row.moveBy(value)).isEqualTo(expected);
+    }
+
+    @Test
+    void moveByException() {
+        Row row = Row.SIX;
+        assertThatThrownBy(() -> row.moveBy(3)).isInstanceOf(IllegalArgumentException.class);
+        assertThatThrownBy(() -> row.moveBy(-6)).isInstanceOf(IllegalArgumentException.class);
+    }
+
+    @Test
+    void getBetween() {
+        Row row = Row.SIX;
+        assertThat(row.getBetween(Row.THREE)).contains(Row.FOUR, Row.FIVE);
+    }
+}
