--- src/test/java/domain/piece/ColorTest.java
@@ -1,13 +1,0 @@
-package domain.piece;
-
-import org.junit.jupiter.api.Test;
-
-import static org.assertj.core.api.Assertions.assertThat;
-
-class ColorTest {
-    @Test
-    void moveUnit() {
-        Color color = Color.BLACK;
-        assertThat(color.moveUnit()).isEqualTo(-1);
-    }
-}
